Move process_expire_borrows to own module.
